Tweak apply the function via a filter hook
