+++ includes/functions.php
@@ -2,8 +2,6 @@
 
 namespace NewfoldLabs\WP\Module\LinkTracker\Functions;
 
-use function NewfoldLabs\WP\ModuleLoader\container as getContainer;
-
 /**
  * Builds a URL with query parameters.
  *
@@ -12,48 +10,9 @@ use function NewfoldLabs\WP\ModuleLoader\container as getContainer;
  * @return string The complete URL with query parameters.
  */
 function build_link( string $url, $params = array() ) {
-
-	$container = getContainer();
-
-	$source = false;
-	if ( ! empty( $params['source'] ) ) {
-		$source = $params['source'];
-		unset( $params['source'] );
-	}
-
-	$parts = wp_parse_url( $url );
-
-	$query_params = array();
-	if ( isset( $parts['query'] ) ) {
-		parse_str( $parts['query'], $query_params );
-	}
-
-	$default_params = array(
-		'channelid'  => strpos( $url, 'wp-admin' ) !== false ? 'P99C100S1N0B3003A151D115E0000V111' : 'P99C100S1N0B3003A151D115E0000V112',
-		'utm_medium' => $container ? $container->plugin()->get( 'id', 'bluehost' ) . '_plugin' : 'bluehost_plugin',
-		'utm_source' => ( $_SERVER['PHP_SELF'] ?? '' ) . ( $source ? '?' . $source : false ),
+	return apply_filters(
+		'nfd_build_url',
+		$url,
+		$params
 	);
-
-	foreach ( $default_params as $key => $value ) {
-		if ( ! array_key_exists( $key, $query_params ) ) {
-			$query_params[ $key ] = $value;
-		}
-	}
-	// Merge the default parameters with the provided parameters and clean the empty parameters.
-	$query_params = array_filter(
-		! empty( $params ) ? array_merge( $params, $query_params ) : $query_params,
-		function ( $value ) {
-			return null !== $value && '' !== $value && false !== $value;
-		}
-	);
-	// Build the final URL with the query parameters.
-	$base = ( isset( $parts['scheme'] ) ? $parts['scheme'] . '://' : '' ) .
-		( isset( $parts['host'] ) ? $parts['host'] : '' ) .
-		( isset( $parts['port'] ) ? ':' . $parts['port'] : '' ) .
-		( isset( $parts['path'] ) ? $parts['path'] : '' );
-
-	// If the original URL has a fragment, append it to the final URL.
-	$fragment = ( isset( $parts['fragment'] ) ? '#' . $parts['fragment'] : '' );
-
-	return $base . '?' . http_build_query( $query_params, '', '&' ) . $fragment;
 }
